requirements.txt
livekit-agents[openai]~=1.0
python-dotenv
mem0ai
pyjwt
flask